refactor(core): move openai types into types/openai/ folder module
