+++ crates/core/src/types/mod.rs
@@ -3,23 +3,23 @@ pub use anthropic::{
     AnthropicSystem, AnthropicTool, AnthropicUsage, DEFAULT_MAX_TOKENS, ThinkingConfig,
 };
 pub use audio::AudioSpeechRequest;
-pub use chat::{
-    ChatCompletionChunk, ChatCompletionRequest, ChatCompletionResponse, Choice, ChunkChoice,
-    CompletionTokensDetails, ContentBlock, Delta, FinishReason, FunctionCall, FunctionCallDelta,
-    FunctionDef, Message, Role, Stop, Tool, ToolCall, ToolCallDelta, ToolChoice, ToolResultContent,
-    ToolType, Usage,
-};
 pub use embedding::{
     Embedding, EmbeddingInput, EmbeddingRequest, EmbeddingResponse, EmbeddingUsage,
 };
 pub use image::ImageRequest;
 pub use model::{Model, ModelList};
 pub use multipart::MultipartField;
+pub use openai::{
+    ChatCompletionChunk, ChatCompletionRequest, ChatCompletionResponse, Choice, ChunkChoice,
+    CompletionTokensDetails, ContentBlock, Delta, FinishReason, FunctionCall, FunctionCallDelta,
+    FunctionDef, Message, Role, Stop, Tool, ToolCall, ToolCallDelta, ToolChoice, ToolResultContent,
+    ToolType, Usage,
+};
 
 mod anthropic;
 mod audio;
-mod chat;
 mod embedding;
 mod image;
 mod model;
 mod multipart;
+mod openai;
