+++ crates/core/src/types/openai/stream.rs
@@ -1,0 +1,101 @@
+use crate::types::openai::{FinishReason, Role, ToolType, Usage};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
+pub struct ChatCompletionChunk {
+    pub id: String,
+    pub object: String,
+    pub created: u64,
+    pub model: String,
+    pub choices: Vec<ChunkChoice>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub usage: Option<Usage>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub system_fingerprint: Option<String>,
+}
+
+impl ChatCompletionChunk {
+    /// Text delta from the first choice, if non-empty.
+    ///
+    /// Empty-string deltas (keepalives, boundary markers) collapse to `None`.
+    pub fn content(&self) -> Option<&str> {
+        self.choices
+            .first()?
+            .delta
+            .content
+            .as_deref()
+            .filter(|s| !s.is_empty())
+    }
+
+    /// Reasoning-content delta from the first choice, if non-empty.
+    ///
+    /// Empty-string deltas collapse to `None`.
+    pub fn reasoning_content(&self) -> Option<&str> {
+        self.choices
+            .first()?
+            .delta
+            .reasoning_content
+            .as_deref()
+            .filter(|s| !s.is_empty())
+    }
+
+    /// Tool-call deltas from the first choice. Empty slice if none.
+    pub fn tool_calls(&self) -> &[ToolCallDelta] {
+        self.choices
+            .first()
+            .and_then(|c| c.delta.tool_calls.as_deref())
+            .unwrap_or(&[])
+    }
+
+    /// Finish reason from the first choice, if present.
+    pub fn finish_reason(&self) -> Option<&FinishReason> {
+        self.choices.first()?.finish_reason.as_ref()
+    }
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
+pub struct ChunkChoice {
+    pub index: u32,
+    pub delta: Delta,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub finish_reason: Option<FinishReason>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub logprobs: Option<serde_json::Value>,
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
+pub struct Delta {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub role: Option<Role>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub content: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub tool_calls: Option<Vec<ToolCallDelta>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub reasoning_content: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
+pub struct ToolCallDelta {
+    pub index: u32,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub id: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "type")]
+    pub kind: Option<ToolType>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub function: Option<FunctionCallDelta>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
+pub struct FunctionCallDelta {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub name: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub arguments: Option<String>,
+}
